Make X509_PUBKEY opaque
